Split up goleveldb files
