+++ goleveldb_test.go
